pyproject.toml
[project]
name = "protobrain"
version = "0.0.3"
description = "Experimental project for machine learning."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.0",
    "pillow>=11.2.1",
    "torch>=2.7.1",
]

[dependency-groups]
dev = [
    "grpcio-tools>=1.73.0",
    "pandas>=2.3.0",
    "plotly>=6.1.2",
    "protobuf>=6.31.1",
    "pytest>=8.4.0",
    "ruff>=0.11.13",
    "streamlit>=1.45.1",
]
